(split)long long replaced by int64_t
git-svn-id: http://svn.xvid.org/trunk@267 060a26f8-e6a2-0410-8959-474133d486e0

+++ src/motion/motion_est.c
@@ -173,7 +173,7 @@ MotionEstimation(MBParam * const pParam,
 
 	const VECTOR zeroMV = { 0, 0 };
 
-	long long time;
+	int64_t time;
 	int32_t x, y;
 	int32_t iIntra = 0;
 	VECTOR pmv;
